Put log CDF for Normal in dist_math for reuse
